LayoutTests/imported/w3c/web-platform-tests/editing/other/copy-paste-in-styled-inline-editing-host.https-expected.txt

ABC DEF GHI

FAIL Copying text in styled inline editing host should not duplicate the editing host assert_equals: expected "ABC DEF GHIDEF " but got "ABC DEF GHI"
